requirements.txt
streamlit
opencv-python
numpy
streamlit_webrtc
streamlit_drawable_canvas
pillow